Moved around classes according to features.
