+++ app/src/main/java/com/esoxjem/movieguide/Api.java
@@ -1,11 +1,11 @@
-package com.esoxjem.movieguide.constants;
+package com.esoxjem.movieguide;
 
 /**
  * @author arun
  */
 public class Api
 {
-    public static final String API_KEY = "";
+    public static final String API_KEY = ""; // add your API key here
     public static final String GET_POPULAR_MOVIES = "http://api.themoviedb.org/3/discover/movie?language=en&sort_by=popularity.desc&api_key=" + API_KEY;
     public static final String GET_HIGHEST_RATED_MOVIES = "http://api.themoviedb.org/3/discover/movie?vote_count.gte=500&language=en&sort_by=vote_average.desc&api_key=" + API_KEY;
     public static final String GET_TRAILERS = "http://api.themoviedb.org/3/movie/%s/videos?api_key=" + API_KEY;
