+++ app/src/main/java/com/esoxjem/movieguide/AppModule.java
@@ -1,4 +1,4 @@
-package com.esoxjem.movieguide.di;
+package com.esoxjem.movieguide;
 
 import android.app.Application;
 import android.content.Context;
