+++ app/src/main/java/com/esoxjem/movieguide/details/IMovieDetailsInteractor.java
@@ -1,7 +1,7 @@
 package com.esoxjem.movieguide.details;
 
-import com.esoxjem.movieguide.entities.Review;
-import com.esoxjem.movieguide.entities.Video;
+import com.esoxjem.movieguide.Review;
+import com.esoxjem.movieguide.Video;
 
 import java.util.List;
