+++ app/src/main/java/com/esoxjem/movieguide/details/IMovieDetailsView.java
@@ -1,8 +1,8 @@
 package com.esoxjem.movieguide.details;
 
-import com.esoxjem.movieguide.entities.Movie;
-import com.esoxjem.movieguide.entities.Review;
-import com.esoxjem.movieguide.entities.Video;
+import com.esoxjem.movieguide.Movie;
+import com.esoxjem.movieguide.Review;
+import com.esoxjem.movieguide.Video;
 
 import java.util.List;
