+++ app/src/main/java/com/esoxjem/movieguide/details/MovieDetailsActivity.java
@@ -6,8 +6,8 @@ import android.support.v7.app.AppCompatActivity;
 import android.view.MenuItem;
 
 import com.esoxjem.movieguide.R;
-import com.esoxjem.movieguide.constants.Constants;
-import com.esoxjem.movieguide.entities.Movie;
+import com.esoxjem.movieguide.Constants;
+import com.esoxjem.movieguide.Movie;
 
 public class MovieDetailsActivity extends AppCompatActivity
 {
