+++ app/src/main/java/com/esoxjem/movieguide/favorites/FavoritesInteractor.java
@@ -1,12 +1,10 @@
 package com.esoxjem.movieguide.favorites;
 
-import com.esoxjem.movieguide.entities.Movie;
+import com.esoxjem.movieguide.Movie;
 
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
-
-import javax.inject.Inject;
 
 /**
  * @author arun
