+++ app/src/main/java/com/esoxjem/movieguide/favorites/FavoritesModule.java
@@ -1,5 +1,6 @@
-package com.esoxjem.movieguide.di;
+package com.esoxjem.movieguide.favorites;
 
+import com.esoxjem.movieguide.AppModule;
 import com.esoxjem.movieguide.favorites.FavoritesInteractor;
 import com.esoxjem.movieguide.favorites.FavoritesStore;
 import com.esoxjem.movieguide.favorites.IFavoritesInteractor;
