+++ app/src/main/java/com/esoxjem/movieguide/listing/IMoviesListingInteractor.java
@@ -1,6 +1,6 @@
 package com.esoxjem.movieguide.listing;
 
-import com.esoxjem.movieguide.entities.Movie;
+import com.esoxjem.movieguide.Movie;
 
 import java.util.List;
