+++ app/src/main/java/com/esoxjem/movieguide/sorting/SortingDialogInteractor.java
@@ -1,7 +1,5 @@
 package com.esoxjem.movieguide.sorting;
 
-import com.esoxjem.movieguide.entities.SortType;
-
 /**
  * @author arun
  */
