+++ app/src/main/java/com/esoxjem/movieguide/sorting/SortingModule.java
@@ -1,10 +1,6 @@
-package com.esoxjem.movieguide.di;
+package com.esoxjem.movieguide.sorting;
 
-import com.esoxjem.movieguide.sorting.ISortingDialogInteractor;
-import com.esoxjem.movieguide.sorting.ISortingDialogPresenter;
-import com.esoxjem.movieguide.sorting.SortingDialogInteractor;
-import com.esoxjem.movieguide.sorting.SortingDialogPresenter;
-import com.esoxjem.movieguide.sorting.SortingOptionStore;
+import com.esoxjem.movieguide.AppModule;
 
 import javax.inject.Singleton;
